Implementando Arquivo Conftest - Nesse tópico você vai aprender como permitir que fixtures sejam reutilizadas por diferentes módulos utilizando o arquivo conftest.py
close #24
